Text.PrettyPrint.GenericPretty

Description

GenericPretty is a haskell library that provides support for automatic derivation of pretty printing functions on user defined data types. The Outputable library is used underneath, the work is done over SDoc types.

The output provided by the library functions is identical to that of Prelude.show, except it has extra whitespace.

For examples of usage please see the README file.

Documentation

pp :: Out a => a -> IO ()

pp is the default Pretty Printer, it uses a line length of 80 and 1.5 ribbons per line (= 53 non-whitespace chars per line) where ribbon is defined as the maximum length of text, excluding whitespace, on a single line

prettyP :: Out a => Int -> Float -> a -> IO ()

prettyP is a partly customizable Pretty Printer It takes the line length and ribbons per line as parameters

prettyStr :: Out a => a -> String

prettyStr returns the result as a string. The returned value is identical to one made by Prelude.show, except for the extra whitespace

fullPP

Arguments

:: Out a

=> a
The value to pretty print

-> PprStyle
The Outputable library style to use (default is defaultUserStyle)

-> Mode
The Pretty library style(mode) to use (default is PageMode)

-> Int
The maximum line length

-> Float
The number of ribbons per line

-> (TextDetails -> b -> b)
Function that handles the text conversion (default is 'outputTxt')

-> b
The end element of the result ( eg: "" or putChar('\n') )

-> b
The pretty printed result

fullPP is a fully customizable Pretty Printer.

outputTxt :: TextDetails -> IO () -> IO ()

outputTxt transforms the text into strings and outputs it directly. This is one example of a function that can handle the text conversion for fullPP.

outputStr :: TextDetails -> String -> String

outputStr just leaves the text as a string. Another example of a function that can handle the text conversion for fullPP.

class Generic a

Representable types of kind *. This class is derivable in GHC with the XDeriveRepresentable flag on.

class Out a where

The class Out is just a wrapper class for Outputable, which passes an extra parameter used to determine when to wrap types up in parentheses

Methods

out :: Int -> a -> SDoc

out is the equivalent of Prelude.showsPrec it generates output identical to show, except for the extra whitespace

outList :: Int -> [a] -> SDoc

outList mimics the behaviour of Prelude.showList used mainly to output strings correctly, and not as lists of characters

genOut :: (Generic a, GOut (Rep a)) => Int -> a -> SDoc

default, generic, out method. User must use this when implementing Out by typing 'out = genOut'